Python: support running tests of Python bits
To run the tests, we need to have an installed Python module,
so install it to a part of the build tree; this leaves a
build/python/quick_der directory that can be added to
PYTHON_PATH even before installing any of the Python modules.

Makes it possible to run tests on a clean system with an out-
of-source build. Adjust the test-template to include that
extra path.

Rename the test-modules, too. Instead of testing <foo>
with a python module named <foo> that does "import <foo>",
call the test-module <foo-test>. In some cases (e.g. running
the test-script from a parent directory) the test-module
<foo> is already registered with sys.modules as <foo>,
so the inner import does nothing -- and so doesn't test the
generated module.