Python: Fix generated RFC modules.
Checking for class-level settings of _context for
types without them (e.g. in rfc4511, which instantiates
an ASN1Integer) triggered an AttributeError on self._context

Since RFC 4511 defines an integer constant, it gets
instantiated; the instantiation needs a context, so
patch up the generator-script to provide context=False
(that's not None) to the concrete instance.