Detected indefinite-length encoding (from BER) and rejecting with EBADMSG
Indefinite-length encoding is used for streams, whose length is not known
at the time of sending.  The header contains 0x80 for the length, and
eventuelly 0x00,0x00 ends the "chunked" transmission.

This is BER, and we try to support that as far as practical, but this is
too far from what Quick DER can support with relative ease.